+++ apps/gecko/src/options/index.tsx
@@ -1,6 +1,6 @@
-import * as ReactDOM from "react-dom";
-import "normalize.css";
-import "./index.css";
 import App from "./App";
+import "normalize.css";
+import * as ReactDOM from "react-dom";
+import "./index.css";
 
 ReactDOM.render(<App />, document.getElementById("root"));
