--- apps/gecko/src/storage/dao/filterList.ts
@@ -1,7 +1,0 @@
-import { DomainList } from "@globalTypes";
-import daoFactory from "./daoFactory";
-
-export const FILTER_LIST_KEY = "filterList";
-export const defaultValue: DomainList = [];
-
-export default daoFactory(FILTER_LIST_KEY, defaultValue);
