--- apps/gecko/src/storage/dao/index.ts
@@ -1,5 +1,0 @@
-import filterList from "./filterList";
-import preferenceList from "./preferenceList";
-import options from "./options";
-
-export { filterList, preferenceList, options };
