--- apps/gecko/src/storage/dao/options.ts
@@ -1,12 +1,0 @@
-import { FilterMode, Options } from "@globalTypes";
-import daoFactory from "./daoFactory";
-
-export const OPTIONS_KEY = "options";
-
-export const defaultValue: Options = {
-  filterListEnabled: true,
-  preferenceListEnabled: true,
-  filterMode: "default",
-};
-
-export default daoFactory(OPTIONS_KEY, defaultValue);
