--- apps/gecko/src/storage/dao/preferenceList.ts
@@ -1,7 +1,0 @@
-import { DomainList } from "@globalTypes";
-import daoFactory from "./daoFactory";
-
-export const PREFERENCE_LIST_KEY = "preferenceList";
-export const defaultValue: DomainList = [];
-
-export default daoFactory(PREFERENCE_LIST_KEY, defaultValue);
