--- apps/gecko/src/storage/index.ts
@@ -1,5 +1,0 @@
-import * as daos from "./dao";
-
-const storage = { ...daos };
-
-export default storage;
