+++ libs/common/storage/index.ts
@@ -1,0 +1,20 @@
+import { DomainList, Options } from "@globalTypes";
+import { storageFactory } from "@utils/storage.util";
+
+// Storage keys and default values
+const filterList = storageFactory("filterList", [] as DomainList);
+const preferenceList = storageFactory("preferenceList", [] as DomainList);
+const options = storageFactory("options", {
+  filterListEnabled: true,
+  preferenceListEnabled: true,
+  filterMode: "default",
+} as Options);
+
+// Storage API
+const storage = {
+  filterList,
+  preferenceList,
+  options,
+};
+
+export { storage };
