+++ libs/common/utils/string.util.ts
@@ -1,0 +1,6 @@
+/**
+ * Escapes a string for safe use within a RegExp constructor.
+ */
+export function escapeRegExp(str: string) {
+  return str.replace(/[.*+?^${}()|[\]\\]/g, "\\$&"); // $& means the whole matched string
+}
